[HAL] Rename apic.h to apicp.h and move it back to the apic folder
