+++ hal/halx86/apic/apic.c
@@ -12,7 +12,7 @@
 /* INCLUDES *******************************************************************/
 
 #include <hal.h>
-#include <apic.h>
+#include "apicp.h"
 #define NDEBUG
 #include <debug.h>
