+++ hal/halx86/apic/apicp.h
